+++ changelog.md
@@ -125,6 +125,9 @@
 
 - Fixed buffer overflow bugs in `net`
 
+- Added `std/cputicks` containing APIs for nanosecond resolution CPU counters, providing strictly monotonic
+  counters with highest available resolution.
+
 - Exported `sslHandle` from `net` and `asyncnet`.
 
 - Added `sections` iterator in `parsecfg`.